fix 429 error sending-9

+++ src/routes/auth-route.ts
@@ -8,8 +8,8 @@ import { mailService } from '../utils/mail-service';
 import { v4 as uuidv4 } from 'uuid';
 import { collections } from '../connect-db';
 import { checkIpServiceLogin, checkIpServiceLogin2 } from '../application/check-Ip-service-login';
-import { checkIpServiceRegistration, checkIpServiceRegistration2 } from '../application/check-Ip-service-registration';
-import { checkIpServiceResending, checkIpServiceResending2 } from '../application/check-Ip-service-resending';
+import { checkIpServiceRegistration2 } from '../application/check-Ip-service-registration';
+import { checkIpServiceResending2 } from '../application/check-Ip-service-resending';
 import { checkIpServiceConfirmation2 } from '../application/check-Ip-service-confirmation';
 import { differenceInSeconds } from 'date-fns';
 
@@ -39,37 +39,44 @@ authRouter.post(
     const result = validationResult(req).formatWith(errorFormatter);
     const userIp = requestIp.getClientIp(req);
     const userConfirmation = await checkIpServiceLogin2.findUserByIp(userIp!);
-    if (!userConfirmation) {
-      await checkIpServiceLogin2.addUser(userIp!);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt < attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      await checkIpServiceLogin2.addAttemptUser(userIp!);
-    }
+
     if (!result.isEmpty()) {
+      if (!userConfirmation) {
+        await checkIpServiceLogin2.addUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation.attempt < attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceLogin2.addAttemptUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation!.attempt >= attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceLogin2.addError429User(userIp!);
+        return res.send(429);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === true
+      ) {
+        await checkIpServiceLogin2.deleteUser(userIp!);
+        return res.send(204);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === false
+      ) {
+        await checkIpServiceLogin2.deleteUser(userIp!);
+      }
+
       return res.status(400).send({ errorsMessages: result.array() });
     }
-    if (userConfirmation && differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit) {
-      // await collections.ipUsers?.updateOne({ userIp }, { $set: { attempt: 1, createdAt: new Date() } });
-      await checkIpServiceLogin2.deleteUser(userIp!);
-      // return next();
-      return res.send(204);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt >= attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      return res.send(429);
-    }
-
-    // const result = validationResult(req).formatWith(errorFormatter);
-    // if (!result.isEmpty()) {
-    //   return res.status(400).send({ errorsMessages: result.array() });
-    // }
     const isCheck = await authRepositoryDB.authUser(req.body.login, req.body.password);
     if (isCheck === 'max limit') {
       await collections.ipUsersLogin?.updateOne({ userIp }, { $set: { attempt: 6 } });
@@ -126,36 +133,47 @@ authRouter.post(
     .withMessage(
       "The field Email must match the regular expression '^[\\\\w-\\\\.]+@([\\\\w-]+\\\\.)+[\\\\w-]{2,4}$'.",
     ),
-  checkIpServiceRegistration,
   async (req, res) => {
     const result = validationResult(req).formatWith(errorFormatter);
     const userIp = requestIp.getClientIp(req);
     const userConfirmation = await checkIpServiceRegistration2.findUserByIp(userIp!);
-    if (!userConfirmation) {
-      await checkIpServiceRegistration2.addUser(userIp!);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt < attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      await checkIpServiceRegistration2.addAttemptUser(userIp!);
-    }
+
     if (!result.isEmpty()) {
+      if (!userConfirmation) {
+        await checkIpServiceRegistration2.addUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation.attempt < attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceRegistration2.addAttemptUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation!.attempt >= attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceRegistration2.addError429User(userIp!);
+        return res.send(429);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === true
+      ) {
+        await checkIpServiceRegistration2.deleteUser(userIp!);
+        return res.send(204);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === false
+      ) {
+        await checkIpServiceRegistration2.deleteUser(userIp!);
+      }
+
       return res.status(400).send({ errorsMessages: result.array() });
-    }
-    if (userConfirmation && differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit) {
-      // await collections.ipUsers?.updateOne({ userIp }, { $set: { attempt: 1, createdAt: new Date() } });
-      await checkIpServiceRegistration2.deleteUser(userIp!);
-      // return next();
-      return res.send(204);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt >= attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      return res.send(429);
     }
 
     // const result = validationResult(req).formatWith(errorFormatter);
@@ -171,7 +189,7 @@ authRouter.post(
       const createdUser = await usersRepositoryDB.deleteUserByLogin(req.body.login);
       return res.status(400).send(createdUser.deleteCount === 1 ? isSendStatus.data : 'failed delete user');
     } else {
-      await checkIpServiceRegistration2.deleteUser(userIp!);
+      // await checkIpServiceRegistration2.deleteUser(userIp!);
       // await collections.ipUsersRegistration?.deleteOne({ userIp });
       return res.status(204).send(isSendStatus.data);
     }
@@ -195,42 +213,49 @@ authRouter.post(
     .withMessage(
       "The field Email must match the regular expression '^[\\\\w-\\\\.]+@([\\\\w-]+\\\\.)+[\\\\w-]{2,4}$'.",
     ),
-  checkIpServiceResending,
+  // checkIpServiceResending,
   async (req, res) => {
     const result = validationResult(req).formatWith(errorFormatter);
     const userIp = requestIp.getClientIp(req);
     const userConfirmation = await checkIpServiceResending2.findUserByIp(userIp!);
-    if (!userConfirmation) {
-      await checkIpServiceResending2.addUser(userIp!);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt < attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      await checkIpServiceResending2.addAttemptUser(userIp!);
-    }
+
     if (!result.isEmpty()) {
+      if (!userConfirmation) {
+        await checkIpServiceResending2.addUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation.attempt < attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceResending2.addAttemptUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation!.attempt >= attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceResending2.addError429User(userIp!);
+        return res.send(429);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === true
+      ) {
+        await checkIpServiceResending2.deleteUser(userIp!);
+        return res.send(204);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === false
+      ) {
+        await checkIpServiceResending2.deleteUser(userIp!);
+      }
+
       return res.status(400).send({ errorsMessages: result.array() });
     }
-    if (userConfirmation && differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit) {
-      // await collections.ipUsers?.updateOne({ userIp }, { $set: { attempt: 1, createdAt: new Date() } });
-      await checkIpServiceResending2.deleteUser(userIp!);
-      // return next();
-      return res.send(204);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt >= attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      return res.send(429);
-    }
-    // const result = validationResult(req).formatWith(errorFormatter);
-    //
-    // if (!result.isEmpty()) {
-    //   return res.status(400).send({ errorsMessages: result.array() });
-    // }
     const newCode = uuidv4();
     await usersRepositoryDB.updateCodeByEmail(req.body.email, newCode);
     const isSendStatus = await mailService.sendEmail(req.body.email, newCode);
@@ -265,31 +290,43 @@ authRouter.post(
     const result = validationResult(req).formatWith(errorFormatter);
     const userIp = requestIp.getClientIp(req);
     const userConfirmation = await checkIpServiceConfirmation2.findUserByIp(userIp!);
-    if (!userConfirmation) {
-      await checkIpServiceConfirmation2.addUser(userIp!);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt < attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      await checkIpServiceConfirmation2.addAttemptUser(userIp!);
-    }
+
     if (!result.isEmpty()) {
+      if (!userConfirmation) {
+        await checkIpServiceConfirmation2.addUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation.attempt < attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceConfirmation2.addAttemptUser(userIp!);
+      }
+      if (
+        userConfirmation &&
+        userConfirmation!.attempt >= attemptsLimit &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
+      ) {
+        await checkIpServiceConfirmation2.addError429User(userIp!);
+        return res.send(429);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === true
+      ) {
+        await checkIpServiceConfirmation2.deleteUser(userIp!);
+        return res.send(204);
+      }
+      if (
+        userConfirmation &&
+        differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit &&
+        userConfirmation!.error429 === false
+      ) {
+        await checkIpServiceConfirmation2.deleteUser(userIp!);
+      }
+
       return res.status(400).send({ errorsMessages: result.array() });
-    }
-    if (userConfirmation && differenceInSeconds(new Date(), userConfirmation!.createdAt) > secondsLimit) {
-      // await collections.ipUsers?.updateOne({ userIp }, { $set: { attempt: 1, createdAt: new Date() } });
-      await checkIpServiceConfirmation2.deleteUser(userIp!);
-      // return next();
-      return res.send(204);
-    }
-    if (
-      userConfirmation &&
-      userConfirmation.attempt >= attemptsLimit &&
-      differenceInSeconds(new Date(), userConfirmation!.createdAt) < secondsLimit
-    ) {
-      return res.send(429);
     } else {
       await checkIpServiceConfirmation2.deleteUser(userIp!);
       return res.send(204);
